services/api_gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.2.1
opensearch-py==2.5.0
httpx==0.27.0
confluent-kafka==2.3.0
